add tests for `baseDir` option
