+++ test/second-entry/constants.js
@@ -1,0 +1,5 @@
+module.exports = {
+  colors: {
+    primary: '#8EE7D3',
+  },
+};
